refactor(flow): inject client interface into the flow
