+++ jina/jaml/parsers/__init__.py
@@ -13,7 +13,7 @@ def _get_all_parser(cls: Type['JAMLCompatible']):
     :return: a tuple of two elements; first is a list of all parsers, second is the legacy parser for default fallback
     """
     from ...executors import BaseExecutor
-    from ...flow import BaseFlow
+    from ...flow.base import BaseFlow
 
     if issubclass(cls, BaseFlow):
         return _get_flow_parser()
