+++ tests/integration/high_order_matches/test_document.py
@@ -1,4 +1,4 @@
-from jina.flow import Flow
+from jina import Flow
 from jina import Document, Executor, requests
