+++ tests/integration/issues/github_2103/test_search_attributes.py
@@ -5,7 +5,7 @@ import time
 import pytest
 from urllib import request
 
-from jina.flow import Flow
+from jina import Flow
 from jina.proto import jina_pb2
 from jina import Document
 from jina import helper
