+++ tests/jinahub/app.py
@@ -1,4 +1,4 @@
-from jina.flow import Flow
+from jina import Flow
 
 if __name__ == '__main__':
     with Flow.load_config('flow.yml') as f:
