+++ tests/unit/flow/test_flow_visualization.py
@@ -1,6 +1,6 @@
 import os
 
-from jina.flow import Flow
+from jina import Flow
 from jina import Executor
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
