+++ tests/unit/flow/test_flow_yaml_parser.py
@@ -4,9 +4,10 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from jina import Flow, AsyncFlow, Executor
+from jina import Executor
+from jina.flow.asyncio import AsyncFlow
 from jina.excepts import BadFlowYAMLVersion
-from jina.flow import BaseFlow
+from jina import Flow
 from jina.jaml import JAML
 from jina.jaml.parsers import get_supported_versions
 from jina.parsers.flow import set_flow_parser
