+++ tests/unit/peapods/runtimes/remote/ssh/test_ssh_remote.py
@@ -1,7 +1,7 @@
 import pytest
 
 from jina.enums import RemoteAccessType
-from jina.flow import Flow
+from jina import Flow
 from jina.parsers import set_pea_parser, set_pod_parser
 from jina.peapods.pods import BasePod
 from jina.peapods.runtimes.ssh import SSHRuntime
